Fix resolve conflict bug

+++ models/experience_like_model.js
@@ -55,7 +55,13 @@ class ExperienceLikeModel {
      *  - reject : DuplicateKeyError/Default Error
      */
     deleteLike(experience_id, user) {
-        return this.collection.deleteOne({
+        const experience_model = new ExperienceModel(this._db);
+        return experience_model.isExist(experience_id).then((is_exist) => {
+            if (!is_exist) {
+                throw new ObjectNotExistError("該篇文章不存在");
+            }
+
+            return this.collection.deleteOne({
                 experience_id: new ObjectId(experience_id),
                 user: user,
             });
